add sprites

+++ Assets/GameZone/Scripts/GameZoneConfig.cs
@@ -8,7 +8,7 @@ namespace GameZone.Scripts
     {
        [field: SerializeField] public int VerticalCount { get; set; }
        [field: SerializeField] public int HorizontalCount { get; set; }
-       [field: SerializeField] public int offset { get; set;}
+       [field: SerializeField] public float offset { get; set;}
        [field: SerializeField] public Vector3 StartPoint { get; set; }
     }
 }
